Bug 482523: disable explicit GC in the idle IDE by default
IDEIdleHelper was introduced in bug 92164 to perform garbage collection
in a way that minimizes impact on responsiveness of the UI.

The boolean system property "ide.gc" was introduced in 3.1 and enabled
by default  (no "ide.gc" needed to enable it) in 3.2 M2 (2005).

This change keeps the IDEIdleHelper job but disables it by default,
unless "ide.gc=true" system property is set.

The reason behind the change:
1) we should let the modern 1.8+ JVM decide when is the best time to
perform GC and finalization
2) and to avoid any unexpected side effects like error popups from
idling UI on OOM reported by the GC.

Change-Id: I3c2d9d1db2d029ab94029fb57011963715a83f7f
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>