Bug 472614 - Smarter import
This wizard and framework allows contributors to contribute a project
detector and configurator. Those detectors are used by this import
framework to automatically identify projects in a given directory. Then
matching projects that have been detected and selected for import are
automatically configured as specified in the contributed class.
Contributors are expected to look at details in the various files to set
as much project properties as possible.
Contributors are expected to be language or framework support, such as
Java support looking for Java file and enabling Java nature and setting
classpath; OSGi support looking for MANIFEST.MF and enabling the
necessary nature...
From user perspective, the operation is only about providing a folder,
reviewing some proposals and hitting finish, and to enjoy project(s)
ready to work with in the IDE.

Bug: 472614
Change-Id: I8ce21fd9fcc696726b1cd082ef01b05bd182826d
Signed-off-by: Mickael Istria <mistria@redhat.com>