Bug 90292: Extension to select how to handle unknown file types
This commit introduces extension point to plug strategies on how to
handle not associated file types. 3 alternative strategies are
contributed by default: "System  editor", "Text editor" and "Ask user
via pop-up" (that last one reuses the "Open With..." dialog).

The default strategy to open editors for unknown file types remains the
"System editor". This strategy first tries to find an in-place editor,
then, if no in-place editor found, tries to find external (system)
editor and if this also cannot be found, uses default text editor.

Another possible strategy, still in the state of an idea and not
contributed here would be something like "Poll a remote
suggestion/discovery service".

The choice between the strategy is managed as a preference, that
RCP providers could set via plugin_customization.ini and/or API; and
that user could set via the "File associations" preference page.

Change-Id: Icd394758a09788a06cba09e17d8951afbf89eb14
Signed-off-by: Mickael Istria <mistria@redhat.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>