Bug 485154 - Properties view reacts on changing to views in the same stack
Properties view should never show content of hidden parts, and parts in
the same stack are "hidden" for the Properties view by definition.

Fix: if the Properties view is hidden, it should not react on activation
of parts in the same stack, because this will invalidate the current
selection (and shown content) anyway.

Change-Id: I2142000d9025ee695c1b8fb8a7186f2385540a84
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>