Bug 417258 - make testSwitchDefaultToExternalBug236104 stable
The test
org.eclipse.ui.tests.api.IEditorRegistryTest.testSwitchDefaultToExternalBug236104
failed because it required a default editor for html to exist. For
whatever reason during the full UI Testsuite run the test fails (most
likely some previously executed test playing with either editor registry
or contributions didn't properly cleaned up), but if executed stand
alone the test passes. 

Since it is hard to find the "bad" test, here we simply do not rely on
html editor anymore - the test is not for html specific editor but for
any existing editor descriptor.

Change-Id: Ib679081f1d96c023c9c154e508156e16c4bef4fc
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>