Bug 417258 - trying to make EvaluationServiceTest stable
The two tests here:
org.eclipse.ui.tests.services.EvaluationServiceTest.testTwoEvaluations
org.eclipse.ui.tests.services.EvaluationServiceTest.testBasicService 

always pass locally BUT always fail if executed in debugger OR Eclipse
window under test is not in focus. So they are depending on the active
window to be *active*, and fail if executed in the environment where the
window is not active for whatever reason (e.g. window manager which pins
other windows on top or disallows focus stealing or simply too slow
window activation code).

The patch tweaks the test timing and tries to force window to become
active, also it makes an assumption about the window state.
Unfortunately, even this assumptions do not work reliably (at least on
GTK2 and KDE) if the Eclipse window under test is manually kept NOT
active during the test.

Change-Id: Idb92efe42afe0f6840c47832711456b700b42cb0
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>