Bug 420479 - [Commands] Add HandlerUtil.getCurrentStructuredSelection
The suggested patch adds a getCurrentStructuredSelection method which
returns an empty collection if a structured selection is not available. 

This is similar to the implementation of StructuredViewer but different
from getCurrentSelection in HandlerUtil which returns null. 

IMHO following the StructuredViewer pattern makes it simpler for
caller,e.g,
org.eclipse.ltk.internal.ui.refactoring.actions.DeleteResourcesHandler.  


Change-Id: I482115731167f18671f696fc8afb2372ea091a62
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>