Bug 438398 - Turning all editors not visible makes the Editors Area not visible as well
Made 2 changes a) to not make Editor area invisible if it is the last
editor area and b) To add or create a CTabItem only if the element is
visible. This is being checked in showTab call but not in the 2 other
functions (processConetnts and childRendered) which call the createTab
function.

Change-Id: I8c7a8c6415480395a753fcc7361c83bd343c9aef
Signed-off-by: Kalyan Prasad <kalyan_prasad@in.ibm.com>