Bug 495139 - [import] - Incorrect import for projects from zip
The condition to decide whether to instantiate and configure a new
import job was failing for archives. So when using an archive as input,
the configuration (directories to import) is set but is then forgotten
by the newer instance of the job.
With the provided fix, the same job is used as long as we do not modify
the path or the import options, so the selection of folders is kept and
used by the import job.


Change-Id: Icd296672c2db8e3d2428fd8a177756db65885dfd
Signed-off-by: Mickael Istria <mistria@redhat.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>