Bug 493391 - Unnatural button order in File Search dialog with GTK3
Rearranged the button order on File search pages so that it reads as
Cancel -> Replace -> Search. Previously, fReplaceButton used
setVisible() to show/hide the button depending on the page, which worked
fine as the order was Replace -> Cancel -> Search, causing no layout
issues. Because Replace is now put in between Cancel and Search, there
was a space left when it was hidden.

Using GridData.exclude() in combination with setVisible() allowed
fReplaceButton to hide and also have the layout ignore the widget,
removing the empty space.

Change-Id: I1978e5e9fe58bcab96b28dfe31df6e4b429cb2cf
Signed-off-by: Ian Pun <ipun@redhat.com>