Bug 495567 - properly handle dirty state changes in properties view
This is the first part of the fix, corresponding to the solution #1
described in bug 495567 comment 6): a part exposing
ISaveablePart either via implementing it or via providing adapter to it,
must properly handle and propagate dirty state changes to the UI.

Properties view will properly track now dirty state of the contributing
ISaveablePart part, also if multiple instances of Properties view are
opened for same contributing part. As a result, UI representation of a
Properties view tracking a ISaveablePart will change: it will show
leading asterisk '*' in all cases the contributing part also shows it,
and hide it otherwise.

Change-Id: I7e2084ad8e250fa78c734ecd2cee6786ce7d9138
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>