Bug 496475 - [IDE] UI deadlock via IDE.getEditorAssociationOverrides()
Root cause for the deadlock: IDE.getEditorAssociationOverrides() calls
into contributed client code holding a class lock. Since it is allowed
to call IDE code from different threads, calling into the potentially
unsafe client code is very dangerous and should be avoided.

The fix avoids deadlock by entirely removing synchronization, and
changing the field access modifier to volatile.

The worst case before the patch: deadlock.
The worst case with the proposed change: multiple creation of
IEditorAssociationOverride instances, only in a very limited time. Since
the extension point and EditorAssociationOverrideDescriptor API have not
specified how many times the
EditorAssociationOverrideDescriptor.createOverride() can be called, it
is better to risk multiple instantiations, because the impact is not as
bad as deadlock.

Change-Id: I0dac96b9c87ad9708590bfc1f298253d93547333
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>