Bug 497051 - Improve text search when multiple files have same location
In case 2 files have the same location, we can assume their content are
the same so search results would be the same.
This change takes advantage of that assumption to reuse search result of
a previous file if it has same location. To take advantage of this
logic, the files to search are sorted by location (so we can simply
compare with previous file rather than storing all search results).

Change-Id: Ied9d7dfc11c9a5b3a83821e9e01d7c6435a1d3a5
Signed-off-by: Mickael Istria <mistria@redhat.com>