Bug 497388 - Fragment position in list tags is not well defined
Refactored the ModelAssembler to use a comparator in order to create a
sorted set of all contributed fragments before the merge. This ensures,
that dependencies between fragments can be resolved. 
The ModelAssemblerFragmentOrderingTests contains test cases for
expectations about the sorting of the set as well as expectation about
the merge result.

Change-Id: I1147199f53ad0bcf4ca53918d15167feb3b528b9
Signed-off-by: Eugen Neufeld <eneufeld@eclipsesource.com>