Bug 500420 - Some perspectives migrated from e3 to e4 have no icon
This is a combination of 2 commits squashed to solve the bug on
R4_6_maintenance.

Below you can see the details of the two committs

Change-Id: I4c832fa0d11618b0c9acd7696cf74bca1733ce67
Message: Added workaround to solve the issue in most common cases. 


Signed-off-by: Patrik Suzzi <psuzzi@gmail.com>

Change-Id: I0e3c6e5306fa66c4dbf404508cadba31c1264fd2
Message: Now, for a custom perspective, we try to get the original id by
considering three different behaviors in saving custom perspective ids: 
[1] keeping spaces (observed in Neon), [2] replacing spaces with
underscores and trimming (in eclipse 3.x according comment 12) [3] just
trimming spaces (observed in Eclipse 3.8)

Change-Id: I3931e6488ac39118acf8a575438e5c8c54eb9027
Signed-off-by: Patrik Suzzi <psuzzi@gmail.com>