Bug 364735 - [Viewers] ViewerComparator violates its general contract
Don't use DecoratingLabelProvider for sorting, see bug 364735 comment
60. The results from this label provider are unstable by design and may
change during the sort operation. Instead, use the original label
provider enclosed by the DecoratingLabelProvider.

A new system property is introduced to restore the old behavior in case
someone relies on the decorations for sorting:
"eclipse.disable.fix.for.bug364735". If this system property is set to
"true", this fix will be disabled.

Change-Id: I67dd03e62b8d32913aacc5670c4b864264d8bc1e
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>