Bug 501997 - Initialize the workspace in a background job
Workspace initialization activates some extensions in a synchronous way
(like IRefreshProvider). The activation of these extensions can take a
long time, so it's better to initialize the full workspace in a
background job.

Note that it does not help the Eclipse workbench to start faster because
EMF asks for ResourcesPlugin#getWorkspace() very early. It blocks until
the workspace is fully initialized, but this does not happen during
bundles initialization.

Change-Id: Ie0b324945315094e9835d367a87016d69f09b8e4
Signed-off-by: Mikael Barbero <mikael@eclipse.org>