Bug 501404 - Workspace save outside of locks
IDEWorkbenchAdvisor.disconnectFromWorkspace() deadlocks if called within
a workspace lock. Deadlock is caused by an attempt of background
ModalContext thread to acquire workspace lock, which is never released
by UI thread waiting for ModalContext to complete. This happened, for
example, when editor was in process of saving a resource.

To prevent deadlock, workspace-bound background operation is postponed
for later under such circumstances.

Change-Id: I27c56c7153f52c2d861af1f3eb6ff1e418c3769e
Signed-off-by: Vasili Gulevich <vasili.gulevich@xored.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>