Bug 492252 - Ensure JobInfo are created with FinishedJobs instance
JobInfo instances were created with a null reference for FinishedJobs in
JobInfoFactory. This change makes the FinishedJobs reference in
ProgressManager non optional, and the JobInfoFactory takes the
ProgressManager's FinishedJobs instance for creation of JobInfos.

This required that FinishedJobs does not require a ProgressManager for
its construction. This was needed to register an
IJobProgressManagerListener to the ProgressManager. Instead of
registering itsels, the ProgressManager now registers the listener. This
resolves the circular dependency between both classes.

On shutdown ProgressManager might already be null. Being resilient for
that.

Change-Id: I7ae3c1f4386f3cf15c23296309599f3e8d5ee19b
Signed-off-by: Karsten Thoms <karsten.thoms@karakun.com>
Also-by: Andrzej Witecki <sodoww@gmail.com>