Bug 251156 - Asynchronous content assist
This provides a new AsyncContentAssistant, performaing computation and
presentation of content assist proposals (and related operations)
asynchronously whenever possible.
Part of this change is also internally allowing the parent
ContentAssistant to support multiple IConteantAssistProcessors (although
this is not exposed through public API). The AsyncContentAssistant does
expose the manipulation of multiple contentAssistProcessors though.
Some fields and methods where also made more visible (usually from
private to protected or package, to maximize the ability to reuse
existing code from the current synchronous framework, in order to make
the feature easier to write, and also to provide a consistent experience
with aysnc and sync content assist.

Change-Id: I3e80c6422eaababc165e955ddeb3539909bfcb9f
Signed-off-by: Mickael Istria <mistria@redhat.com>