Bug 508498: Add "information" color api
Introducing a color to be used by widgets that display text.
This is to be used by various Black/White color related bugs in JFace on
Gtk.
This gives us the flexibility to easily change the underlying system
color later when/if COLOR_TEXT_* is implemented. (See comment).

Patch set 6 notes:
Note: This is compatible with theaming. On Gtk dark theme it is dark, on
bright theme it is bright. In fact HOVER_ uses the same mechanism.
Further, this color is not in conflict with higher level theming, 
higher level themes should override system defaults with their own
colors if need be. The goal of this color is to be a suitable substitue
'default' color to be used instead of COLOR_INFO_*

Change-Id: Ib5a5f749359cc0743fd68a5313b1b9a6622c8497
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=508498
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>