Bug 507298: Use fitting color for Heap Status
Heapstatus is not a tooltip. It should not be using tooltip colors.
This improper use of Tooltip color makes heapstatus have very
poor contrast on Gtk. (See screenshot)
Since it is a regular widget, it should be using regular widget
colors.
(See screenshot attached to bug to see before/after).

Change-Id: I18962495557c6e6742b4ce8af6dc51c8c7a43ddb
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507298
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>