Bug 508787 - Rework launch configuration selection dialog for groups
According to discussion in the bug 508787, a better layout for the
selection dialog has been implemented. This also removed quite some
(invisible) widgets from the dialog and should improve the overal
experience with it.

The most user visible change is that the launch mode combo is now below
the launch configuration selection tree.

Change-Id: I25296f5a4c9315af87d817c164a3d33b858919c2
Signed-off-by: Markus Duft <markus.duft@ssi-schaefer.com>