Bug 502841 - TreeViewer::expandAll() causes cursor to flicker
createChildren(), changes the cursor from arrow to hour-glass and back
while TreeItems are created. In some cases, like e.g. expandAll(), this
method is call frequently (here for all tree items) and causes unwanted
flicker.

Move BusyIndicator::showWhile() to handleTreeExpand() where it is only
shown if a user causes the tree item to expand.

This leaves the application code to decide whether some busy indicator
is shown or not and to surround multiple tree operations with a single
busy indicator.

Change-Id: I42a7c6826a6e2afae54b8db2868576ffc54265ad
Signed-off-by: Rüdiger Herrmann <ruediger.herrmann@gmx.de>