Bug 514319 - Using a dedicated trim method to trim the expressions labels
When selected expressions get copied into the clipboards text transfer,
first cut off the leading and tailing quotes.
This is done by extending the VirtualCopyToClipboardActionDelegate with
a method called trimLabel. Clients (in that case a own action delegate
for expressions) could do additional trimming to the label provided by
the label provider.

Change-Id: I71a5c205f7eb2a8db5d112eec2d3a28dbb237f1e
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>