Bug 516137 - Context menu in Ctrl+E is gone and right-click wrongly
Now "Delete" key closes the selected editor. Also the right-click on an
editor element opens a "Close" menu which allows to close the selected
editor. Added saving of the dialog with bypass of focus lost, needed in
the case the user is requested whether to save a dirty editor which is
closing.

Change-Id: I525d13d3ee7c8159d933afdcf4c6fb2ba190c08a
Signed-off-by: Patrik Suzzi <psuzzi@gmail.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>