Bug 517068 - key binding conflicts check for enabled commands
Invoking a keybinding wasn't taking into account whether command is
executable or not. This can be annoying in case of multiple commands
bound to the same binding: user got asked to choose which command to
invoke, but some of those commands were not executable and resulted in
an error message to user.
With a filter to only use executable commands, in most cases, we
shouldn't even see a pop-up as commands on a same shortcut are usually
mutually exclusive, and in case the pop-up shows up, it would only
propose commands that can run properly.

Change-Id: Ic62cd279652fd8ed42c58b66347e3dafea3861e4
Signed-off-by: Mickael Istria <mistria@redhat.com>