Bug 517068 - key binding conflicts check for enabled commands
Invoking a keybinding wasn't taking into account whether command is
executable or not. This can be annoying in case of multiple commands
bound to the same binding: user got asked to choose which command to
invoke, but some of those commands were not executable and resulted in
an error message to user.
With a filter to only use executable commands, in most cases, we
shouldn't even see a pop-up as commands on a same shortcut are usually
mutually exclusive, and in case the pop-up shows up, it would only
propose commands that can run properly.

Cherry-picked from 3b4b24c0ed0ab0ea759805b35222bc72837b8234

Change-Id: Ic9d1decb505240e726f9faa067f4a1ecf6d1556e
Signed-off-by: Mickael Istria <mistria@redhat.com>