Bug 519815 - Associate Editor with ContentType from Preferences
Added new viewer with respective buttons to do the actual associations.
It allows end-users to understand where a file/editor association is 
provided. Currently, only File/editor mapping are shown (and editable) 
as preferences, and user don't get a chance to understand what defines 
other (content-type based) associations.
As user can define content-types, the next natural action they'd like to
do once the content-type is defined is to associate an editor with it. 
At the moment, defining content-type still requires to use a File/Editor
association to associate new types with editors

Change-Id: Ie6be642f560de0b01f08d1238c97a97ad4f0072c
Signed-off-by: Mickael Istria <mistria@redhat.com>