Bug 519934 - [Generic Editor] allow stateful extensions
Currently, the way extension points and registry are implemented and
used make that the same extension (ie same PresentationReconciler) is
used for all instances of the Generic Editor that match the
content-type.
As all known extensions so far manage to be stateless and as most APIs
are designed to not need the class to keep state, this wasn't much of an
issue.
However, the introduction of reconciler and some use-cases related to it
have highlighted that some classes need to be stateful, and that there
is nothing that  enforce them be be stateless. So we need the Generic
Editor to assume by default that classes are stateful, and to provide a
distinct instance of the extension for each editor instance.


Change-Id: Ibfabd88685820df81fe5c53dbe754bb34e2b521e
Signed-off-by: Mickael Istria <mistria@redhat.com>