Bug 520069 - Better performance when setting viewer filters
Suppress updating item labels when filters of a viewer are updated,
since updating the filters will not modify any of the existing labels.
This can reduce the refresh() time drastically.

The change in AbstractTreeViewer seems to be an independent bug, which
was only uncovered by this change. If labelRefreshing is disabled, the
association between viewer elements and widgets will get out of sync,
but that execution path was probably not exploited up to now.

Change-Id: Ic94999d9bcba49a94d2157b2e50cc411e4f9c850
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>