Bug 508508 - Browser editor to allow auto-refresh on resource change
Added feature to BrowserViewer that can toggle auto-refresh.
Auto-refresh is implemented through a WatchService, which will monitor
the html file that is currently being displayed. Any changes to the code
will trigger a browser refresh, showing the newest changes. The feature
is under the dropdown arrow of the current Refresh button.

Change-Id: I4e8ce862cfdc846fe2e776abe757857932f6ff99
Signed-off-by: Lucas Bullen <lbullen@redhat.com>