Bug 521490 - Handle RESOURCE_NOT_FOUND error
When the resource is deleted on the file system and out of synch with
the workspace, a ResourceException is raised while trying to load the
contents. For this case no error has to be raised and the method should
return gracefully like in the previous check for existence of the
resource.

Change-Id: Ic37c767ae38d665ecf4a47a1f4da7d5decd1dcd6
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>