Bug 268608 - find/replace action disabled on first opening of console
Whenever a new console page is opened, a FindReplaceAction is registered
for this console page. The target of this action is set before the new
console page is set as active.

This results in either a disabled find/edit action, if there was no
previous console page, or an action which targets the previously active
console page. I.e. the user either cannot search in the new console
page, or searches in an old one.

This change ensures that the created FindReplaceAction has the correct
target page: the target page is updated once the console page is
activated by the ConsoleView. This fixes the bug.

Change-Id: Ib59be79b361b91949e7ace4f4e1803b4039cad95
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>