Bug 517411 - Race condition between Autobuild and explicit Project build
Running a project build from multiple jobs can result in exceptions in
BuildCommand, since a builder will be added twice for the same
BuildCommand. This sometimes hinders running a project build while
auto-build is on, e.g. in automated tests.

This change ensures two things. First that BuildCommand has synchronous
access to its store for builders. And second, that
BuildManager.getBuilder can deal with situations where another builder
was added to the same BuildCommand in parallel.

Change-Id: I272768ed84a9453f6360c618523c465b06d5d591
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>