Bug 522241 - org.eclipse.ui.console mixes tabs and spaces code indent
The source of org.eclipse.ui.console has a mix of tabs and spaces, in
some caes on the same line. This makes editing the source very awkward.
E.g. changing code and undo-ing changes results in indentation changes
nevertheless (on save action).

In this change space indentations are replaced with tab indentations,
for the source of plug-in org.eclipse.ui.console. This follows the
project formatting settings.

Change-Id: I9ee197b4eeea1c2764b2e55dce55bbeb3881a64a
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>