Bug 528150 - BuildManager silently skip builds running from != threads.
The BuildManager used to have some "canRun" and a "building" guards
method and field which was preventing it from being able to run multiple
independent (with no conflicting scheduling rules) builds at the same
time. The effect was moreover silent: the build was simply skipped and
returning OK, just like if it had concluded successfully, whereas it
actually didn't do anything (and should return such piece of information
or throw an exception).

Change-Id: Idce4239b176b6cf077399e2c1638e55ed6073903
Signed-off-by: Mickael Istria <mistria@redhat.com>