528345 - [Generic Editor] Move Toggle Highlight button to "Editor Presentation"
Contribute "Toggle Highlight" button to "Editor Presentation" instead of
a dedicated toolbar section

For some reason, doing that also makes the hidden separator used as a
group marker for the "Presentation" group visible, so make it invisible
again via an additional contribution. And make sure to do that separately,
because although it's always present, it can't be addressed in a menuContribution otherwise for some reason.

Change-Id: Ie5dab5d45ea420e3af81a2b370fc2a40cd2dbb89
Signed-off-by: Carsten Reckord <reckord@yatta.de>