Bug 528720 - View toolbar painted over view when switching from editor
When an editor tab and a view tab are in the same part stack, switching
from the editor to the view can result in a toolbar painting artifact.
Namely, the view toolbar is painted over the view client area. This can
be observed if the part stack tab area is not long enough to hold the
view toolbar.

This is the case, since the toolbar is drawn within the view area, on
top of the view. The view does not receive a layout update notification,
due to switching from the editor. When switching from another view,
StackRenderer.adjustTopRight will hide the toolbar of the view. This
results in a layout update of the newly opened view. Since the editor
has no toolbar, nothing is hidden and no layout update is queued.

This change improves on the fix for Bug 461573. Instead of triggering a
direct layout update, we request one later on. This later on includes
the adjustment of the toolbar location, resulting in a correctly sized
view. The view toolbar is then no longer painted over the view.

Change-Id: If23c1ed2ab1fe8edc052d647dd89ebed1ed00a0a
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>