Bug 517393: Dark theme hyper-link text is dark blue
Dark theme hyper-link text was dark blue on black background in JDT's
Javadoc. The cause is that JDT's CSS contains hard-coded colors for
the links.
To solve this provide a mechanism in HTMLPrinter to replace symbolic
color names with actual color values from the preferences.

A consumer of HTMLPrinter now can you the symbolic names
hyperlinkColor and
activeHyperlinkColor in their stylesheets.
These symbolic names are replaced with
JFacePreferences.HYPERLINK_COLOR and
JFacePreferences.ACTIVE_HYPERLINK_COLOR by HTMLPrinter automatically.

Also fixed: If one changed the theme or color preferences the cached
colors in HTMLPrinter were not updated.

Change-Id: I386c0a7feff6da132b76bd3ff80c91b4fb919b67
Signed-off-by: Matthias Becker <ma.becker@sap.com>