Bug 371586 - various smaller code fixes in Markers area
- Add toString() to MarkerCategory and MarkerField
- Use Linked* collections for stable iteration behavior
- faster compare in MarkerComparator
- mark fields accessed from different threads volatile
- trying to avoid concurrent access to markerEntryArray:
Markers.getClone() should return cloned arrays

- Note: getClone() implementation and its usage overall in the code is
MT broken by design. It tries to return new instance with same
markerEntryArray object if it is not "inChange" yet. However no one
prevents the change of a "clone" AFTER the array reference was returned,
and no one guarantees the atomicity of the isInChange/markerEntryArray
states.

Change-Id: I439a8e40e0d0302bfd82915cbb23081e0cac2c90
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>